chore(rust/op-reth): port historical proofs binary change (#19252)
* feat: add binary entry point for external proofs in OP (op-rs/op-reth#222)

Closes op-rs/op-reth#164

---------


* feat: add support for `eth_getProof` (op-rs/op-reth#257)

Adds support for `eth_getProof` RPC method. This required reworking the
launch command to still work with an alternative DB provider.

Closes op-rs/op-reth#173

---------


* perf(trie): feature gate `reth-optimism-trie` metrics (op-rs/op-reth#282)

Closes https://github.com/op-rs/bin/issues/281

- Feature gates `reth-optimism-trie` metrics
- Moves cursor impls out of proofs module into new module `cursor`
- Moves cursor factory impls into new module `cursor_factory`
- Updates cursor factory impls to return cursor types with metrics
wrapper if metrics feature is enabled

* fix(test): Enable live collector tests with metrics feature (op-rs/op-reth#291)

Closes https://github.com/op-rs/bin/issues/283

Enable live collector tests when metrics feature is enabled

* feat: implement `debug_executePayload` (op-rs/op-reth#276)

Closes https://github.com/op-rs/bin/issues/189

---------


* feat: live collector integration (op-rs/op-reth#306)

Closes op-rs/op-reth#296

---------


* fix: rebase conflicts (op-rs/op-reth#367)

Fix conflicts rebasing onto latest upstream main

* feat: implemented `OpProofStorage` Database metrics (op-rs/op-reth#407)

Closes op-rs/op-reth#224
Closes op-rs/op-reth#387

---------


* feat: Implemented `OpProofStoragePrunerTask` (op-rs/op-reth#375)

Closes op-rs/op-reth#361
Closes op-rs/op-reth#395

---------


* chore: inmem proof storage removed (op-rs/op-reth#465)

This PR removes the CLI support for in-memory proof storage
(proofs_history_storage_in_mem). The in-memory storage implementation
was primarily intended for unit testing and is not feasible for running
a node due to the large memory requirements of storing historical trie
nodes.

Closes op-rs/op-reth#466

* chore(exex): Add metrics feature in `reth-optimism-exex` (op-rs/op-reth#438)

Closes https://github.com/op-rs/bin/issues/427

---------


* feat: prune cli added (op-rs/op-reth#507)

Closes op-rs/op-reth#452

* refactor(trie): return `OpProofsStorageError` from `execute_and_store_block_updates` (op-rs/op-reth#535)

Closes op-rs/op-reth#523

---------


* fix: reduce default proofs pruning interval (op-rs/op-reth#560)

Closes op-rs/op-reth#559

* feat: add verification interval for integrity check (op-rs/op-reth#577)

Closes op-rs/op-reth#449

The approach used is to perform full block verification after every N
blocks to ensure the state is still correct.

---------


* chore: getProof benchmark utility added (op-rs/op-reth#550)

Utility for op-rs/op-reth#446

* chore: mv proof args to rollup node (op-rs/op-reth#625)

Closes op-rs/op-reth#613

* chore: ExEx config builder (op-rs/op-reth#642)

Closes op-rs/op-reth#641

---------


* chore: moved proof initialization to `reth-optimism-node` (op-rs/op-reth#640)

Closes op-rs/op-reth#612

---------


* chore(rust): fix compilation errors

* docs(op-reth): add historical proofs README to exex crate


* chore(op-reth): move proof-bench to rust/op-reth/bin/proof-bench


* fix(op-reth): fix go-lint and rust-clippy CI failures

- Replace big.Int.Uint64() calls with bigs.Uint64Strict() in proof tests
- Fix goimports ordering in proof test files (move ethereum-optimism/* before ethereum/go-ethereum/*)
- Fix const alignment in preset.go
- Fix uninlined_format_args clippy lint in proof-bench report.rs


fix(op-reth): update proofs preset to use stack.ComponentID

Replace removed type aliases (L1NetworkID, L1ELNodeID, L2CLNodeID, etc.)
with unified stack.ComponentID type. Also fix L2NodeMatcher generics and
match.Not calls to use stack.ByID wrapper.

fix(op-reth): remove duplicate crates/tests/proofs preset

This was a partial copy of rust/op-reth/tests/proofs/utils with no
go.mod, no test files, and no other sources. It caused go-lint CI
failures due to stale type references.

---------
