+++ rust/op-reth/bin/Cargo.toml
@@ -19,9 +19,12 @@ reth-optimism-evm.workspace = true
 reth-optimism-payload-builder.workspace = true
 reth-optimism-primitives.workspace = true
 reth-optimism-forks.workspace = true
+reth-node-builder.workspace = true
+reth-db.workspace = true
 
 clap = { workspace = true, features = ["derive", "env"] }
 tracing.workspace = true
+eyre.workspace = true
 
 [lints]
 workspace = true
@@ -33,6 +36,7 @@ otlp = ["reth-optimism-cli/otlp"]
 
 js-tracer = [
     "reth-optimism-node/js-tracer",
+    "reth-node-builder/js-tracer",
 ]
 
 jemalloc = ["reth-cli-util/jemalloc", "reth-optimism-cli/jemalloc"]
