+++ rust/op-reth/bin/proof-bench/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "op-reth-proof-bench"
+version = "1.11.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+# reth
+reth-cli-runner.workspace = true
+
+# alloy
+alloy-primitives = { workspace = true, features = ["serde"] }
+
+tokio = { workspace = true, features = ["macros"] }
+reqwest = { workspace = true, features = ["json"] }
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+clap = { workspace = true, features = ["derive"] }
+futures.workspace = true
+anyhow.workspace = true
+hdrhistogram = "7.5"
