+++ rust/op-reth/bin/src/main.rs
@@ -1,8 +1,12 @@
 #![allow(missing_docs, rustdoc::missing_crate_level_docs)]
 
 use clap::Parser;
+use eyre::ErrReport;
+use reth_db::DatabaseEnv;
+use reth_node_builder::{NodeBuilder, WithLaunchContext};
+use reth_optimism_chainspec::OpChainSpec;
 use reth_optimism_cli::{Cli, chainspec::OpChainSpecParser};
-use reth_optimism_node::{OpNode, args::RollupArgs};
+use reth_optimism_node::{args::RollupArgs, proof_history};
 use tracing::info;
 
 #[global_allocator]
@@ -11,6 +15,17 @@ static ALLOC: reth_cli_util::allocator::Allocator = reth_cli_util::allocator::ne
 #[cfg(all(feature = "jemalloc-prof", unix))]
 #[unsafe(export_name = "_rjem_malloc_conf")]
 static MALLOC_CONF: &[u8] = b"prof:true,prof_active:true,lg_prof_sample:19\0";
+
+/// Single entry that handles:
+/// - no proofs history (plain node),
+/// - in-mem proofs storage,
+/// - MDBX proofs storage.
+async fn launch_node(
+    builder: WithLaunchContext<NodeBuilder<DatabaseEnv, OpChainSpec>>,
+    args: RollupArgs,
+) -> eyre::Result<(), ErrReport> {
+    proof_history::launch_node_with_proof_history(builder, args).await
+}
 
 fn main() {
     reth_cli_util::sigsegv_handler::install();
@@ -23,11 +38,10 @@ fn main() {
     }
 
     if let Err(err) =
-        Cli::<OpChainSpecParser, RollupArgs>::parse().run(async move |builder, rollup_args| {
+        Cli::<OpChainSpecParser, RollupArgs>::parse().run(async move |builder, args| {
             info!(target: "reth::cli", "Launching node");
-            let handle =
-                builder.node(OpNode::new(rollup_args)).launch_with_debug_capabilities().await?;
-            handle.node_exit_future.await
+            launch_node(builder, args.clone()).await?;
+            Ok(())
         })
     {
         eprintln!("Error: {err:?}");
