--- rust/op-reth/crates/tests/proofs/contracts/foundry.toml
@@ -1,3 +1,0 @@
-[profile.default]
-src = "src"
-out = "artifacts"
